Remove provisioning steps
